requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
faster-whisper==1.1.1
sentence-transformers==3.0.1
faiss-cpu==1.13.2
transformers==4.55.0
torch>=2.3.0
torchaudio>=2.3.0
torchvision>=0.18.0
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
yt-dlp==2025.1.26
ffmpeg-python==0.2.0
jinja2==3.1.6
httpx==0.27.2

# Vector DB
qdrant-client==1.12.1

# LlamaIndex (retrieval + synthesis)
llama-index==0.11.23
llama-index-vector-stores-faiss==0.2.1
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-ollama==0.3.3
